clean up index
Moved endpoint callbacks to their own files because index.js was getting too
long with all of them there. Also moved randName() and returnError() into
constants.js
