+++ endpoints.js
@@ -1,0 +1,13 @@
+const postUpload = require("./endpoints/postUpload");
+const getTour = require("./endpoints/getTour");
+const getEdit = require("./endpoints/getEdit");
+const postEdit = require("./endpoints/postEdit");
+const getTours = require("./endpoints/getTours");
+
+module.exports = {
+  postUpload,
+  getTour,
+  getEdit,
+  postEdit,
+  getTours
+};
